define "bigint" to "long long int" for consistenty
this will silence a lot of warning about improper formats
and alike, since on x86_64 for example int64_t is typedef'd
to "long int" but we assume it to be "long long int" everywhere.

+++ src/lmptype.h
@@ -47,7 +47,8 @@ namespace LAMMPS_NS {
 
 typedef int smallint;
 typedef int tagint;
-typedef int64_t bigint;
+//typedef int64_t bigint;
+typedef long long int bigint;
 
 #define MAXSMALLINT INT_MAX
 #define MAXTAGINT INT_MAX
